+++ backend/tests/__init__.py
@@ -1,0 +1,1 @@
+import backend.tests.default.nodeTypes as default
